refactor: 시험 상태 컬럼 추가

+++ app/controllers/api/v1/exams_controller.rb
@@ -2,9 +2,10 @@
 
 class Api::V1::ExamsController < ApplicationController
   before_action :check_admin, only: [:create, :destroy, :update]
+  before_action :set_exam, only: [:update, :destroy, :show]
 
   def index
-    @exams = Exam.all.page(params[:page] || 1).per(params[:per_page] || 10).map do |exam|
+    @exams = Exam.where(status: Exam::Status::ACTIVATED).page(params[:page] || 1).per(params[:per_page] || 10).map do |exam|
       exam_json = exam.as_json.as_json.deep_transform_keys { |key| key.to_sym }
       exam_json[:count] = Booking.where(exam_id: exam.exam_id, status: Booking::Status::CONFIRMED).count
       exam_json[:id] = exam_json[:exam_id]
@@ -13,6 +14,10 @@ class Api::V1::ExamsController < ApplicationController
     end
 
     render_json_response(data: @exams)
+  end
+
+  def show
+    render_json_response(data: delete_primary_key(@exam))
   end
 
   def create
@@ -26,8 +31,6 @@ class Api::V1::ExamsController < ApplicationController
   end
 
   def update
-    @exam = Exam.find(params[:id])
-
     if @exam.update(exam_params)
       render_json_response(data: delete_primary_key(@exam))
     else
@@ -36,13 +39,19 @@ class Api::V1::ExamsController < ApplicationController
   end
 
   def destroy
-    @exam = Exam.find(params[:id])
-    @exam.destroy
-
-    render_json_response(data: delete_primary_key(@exam))
+    if @exam.set_delete
+      render_json_response(data: delete_primary_key(@exam))
+    else
+      render_json_response(message: @exam.errors.full_messages.join(', '), status: 422)
+    end
   end
 
   private
+
+  def set_exam
+    @exam = Exam.find_by(exam_id: params[:id], status: Exam::Status::ACTIVATED)
+    render_json_response(message: 'Exam not found.', status: 404) unless @exam
+  end
 
   def delete_primary_key(exam)
     exam_json = exam.as_json.deep_transform_keys { |key| key.to_sym }
